cavali-service-5/requirements.txt
# ---- Servidor Web ----
fastapi
uvicorn[standard]
python-dotenv
google-cloud
requests
sqlalchemy
pydantic
google-cloud-storage
google-cloud-pubsub